feat: T-986 Edit / Delete billable metric permissions (#67)

+++ app/services/billable_metrics_service.rb
@@ -7,7 +7,7 @@ class BillableMetricsService < BaseService
       name: args[:name],
       code: args[:code],
       description: args[:description],
-      aggregation_type: args[:aggregation_type]&.to_sym
+      aggregation_type: args[:aggregation_type]&.to_sym,
     )
 
     result.billable_metric = metric
@@ -21,9 +21,15 @@ class BillableMetricsService < BaseService
     return result.fail!('not_found') unless metric
 
     metric.name = args[:name]
-    metric.code = args[:code]
     metric.description = args[:description] if args[:description]
-    metric.aggregation_type = args[:aggregation_type]&.to_sym
+
+    # NOTE: Only name and description are editable if billable metric
+    #       is attached to subscriptions
+    unless metric.attached_to_subscriptions?
+      metric.code = args[:code]
+      metric.aggregation_type = args[:aggregation_type]&.to_sym
+    end
+
     metric.save!
 
     result.billable_metric = metric
@@ -35,6 +41,13 @@ class BillableMetricsService < BaseService
   def destroy(id)
     metric = result.user.billable_metrics.find_by(id: id)
     return result.fail!('not_found') unless metric
+
+    unless metric.deletable?
+      return result.fail!(
+        'forbidden',
+        'Billable metric is attached to an active subscriptions',
+      )
+    end
 
     metric.destroy!
